add `GenericError`
